+++ packages/ui/src/core/Paper.tsx
@@ -3,7 +3,7 @@ import { Paper as MuiPaper, PaperProps as MuiPaperProps } from '@mui/material'
 
 export interface PaperProps extends MuiPaperProps {}
 
-const Paper: React.FC<PaperProps> = props => {
+const Paper: React.FC<PaperProps> = (props) => {
   return <MuiPaper {...props} />
 }